Added orphan remover and refined the logging and the dry-run mode

+++ src/features/synchronizer.ts
@@ -5,6 +5,7 @@
 class Synchronizer {
 
     private readonly logger;
+    private readonly print;
 
     //------------------------------------------------------------------------------------------------------------------
     // Initialization
@@ -12,6 +13,7 @@ class Synchronizer {
 
     public constructor(private readonly context: Context) {
         this.logger = context.logger;
+        this.print = context.print;
     }
 
     //------------------------------------------------------------------------------------------------------------------
@@ -22,12 +24,11 @@ class Synchronizer {
         this.logger.info(this.context.options.dryRun ? "Simulating synchronization" : "Starting synchronization");
         const json = JsonLoader.loadAndValidateDatabase(this.context);
         const database = new DatabaseAssembler(this.context).assembleFromJson(json);
-        // Delete files in the destination that are not linked in the database
+        new OrphanRemover(this.context).run(database);
         // Copy over all missing files from the source
         // Create the recovery/index files
         this.saveDatabase(database);
     }
-
 
     //------------------------------------------------------------------------------------------------------------------
     // Save the database
@@ -35,7 +36,13 @@ class Synchronizer {
 
     private saveDatabase(database: MappedRootDirectory) {
         const file = this.context.files.database;
-        this.logger.info(`Saving database ${file}`);
-        node.fs.writeFileSync(file, DatabaseSerializer.serialize(database));
+        if (this.context.options.dryRun) {
+            this.logger.info(`Would save database ${file}`);
+            this.print("Would save updated database");
+        } else {
+            this.logger.info(`Saving updated database ${file}`);
+            this.print("Saving updated database");
+            node.fs.writeFileSync(file, DatabaseSerializer.serialize(database));
+        }
     }
 }
